LPD-58182 Inline MatchQueryTranslator/MatchQueryTranslatorImpl
