+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/query/ElasticsearchQueryTranslator.java
@@ -6,6 +6,7 @@
 package com.liferay.portal.search.elasticsearch7.internal.query;
 
 import com.liferay.petra.function.transform.TransformUtil;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -71,6 +72,7 @@ import java.util.TimeZone;
 
 import org.elasticsearch.common.bytes.BytesArray;
 import org.elasticsearch.common.bytes.BytesReference;
+import org.elasticsearch.common.unit.Fuzziness;
 import org.elasticsearch.index.query.BoostingQueryBuilder;
 import org.elasticsearch.index.query.CommonTermsQueryBuilder;
 import org.elasticsearch.index.query.GeoBoundingBoxQueryBuilder;
@@ -80,6 +82,7 @@ import org.elasticsearch.index.query.GeoShapeQueryBuilder;
 import org.elasticsearch.index.query.IdsQueryBuilder;
 import org.elasticsearch.index.query.MatchPhrasePrefixQueryBuilder;
 import org.elasticsearch.index.query.MatchPhraseQueryBuilder;
+import org.elasticsearch.index.query.MatchQueryBuilder;
 import org.elasticsearch.index.query.PrefixQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
@@ -88,6 +91,7 @@ import org.elasticsearch.index.query.RegexpQueryBuilder;
 import org.elasticsearch.index.query.SimpleQueryStringBuilder;
 import org.elasticsearch.index.query.TermsSetQueryBuilder;
 import org.elasticsearch.index.query.WildcardQueryBuilder;
+import org.elasticsearch.index.query.ZeroTermsQueryOption;
 import org.elasticsearch.index.query.functionscore.FunctionScoreQueryBuilder;
 import org.elasticsearch.index.query.functionscore.ScoreFunctionBuilder;
 import org.elasticsearch.legacygeo.builders.ShapeBuilder;
@@ -485,8 +489,7 @@ public class ElasticsearchQueryTranslator
 
 	@Override
 	public QueryBuilder visit(MatchQuery matchQuery) {
-		return _addBoost(
-			matchQuery, _matchQueryTranslator.translate(matchQuery));
+		return _addBoost(matchQuery, _translate(matchQuery));
 	}
 
 	@Override
@@ -758,6 +761,44 @@ public class ElasticsearchQueryTranslator
 			"Invalid FunctionScoreQuery.ScoreMode: " + scoreMode);
 	}
 
+	private QueryBuilder _translate(MatchQuery matchQuery) {
+		String field = matchQuery.getField();
+
+		MatchQuery.Type type = matchQuery.getType();
+		Object value = matchQuery.getValue();
+
+		if (value instanceof String) {
+			String stringValue = (String)value;
+
+			if (stringValue.startsWith(StringPool.QUOTE) &&
+				stringValue.endsWith(StringPool.QUOTE)) {
+
+				type = MatchQuery.Type.PHRASE;
+
+				stringValue = StringUtil.unquote(stringValue);
+
+				if (stringValue.endsWith(StringPool.STAR)) {
+					type = MatchQuery.Type.PHRASE_PREFIX;
+				}
+			}
+
+			if (type == MatchQuery.Type.PHRASE) {
+				return _translateMatchPhraseQuery(
+					field, stringValue, matchQuery);
+			}
+			else if (type == MatchQuery.Type.PHRASE_PREFIX) {
+				return _translateMatchPhrasePrefixQuery(
+					field, stringValue, matchQuery);
+			}
+		}
+
+		if ((type == null) || (type == MatchQuery.Type.BOOLEAN)) {
+			return _translateMatchQuery(field, value, matchQuery);
+		}
+
+		throw new IllegalArgumentException("Invalid match query type: " + type);
+	}
+
 	private org.elasticsearch.index.query.Operator _translate(
 		Operator matchQueryOperator) {
 
@@ -828,6 +869,111 @@ public class ElasticsearchQueryTranslator
 			scoreFunctionBuilder);
 	}
 
+	private QueryBuilder _translateMatchPhrasePrefixQuery(
+		String field, String value, MatchQuery matchQuery) {
+
+		MatchPhrasePrefixQueryBuilder matchPhrasePrefixQueryBuilder =
+			QueryBuilders.matchPhrasePrefixQuery(field, value);
+
+		if (Validator.isNotNull(matchQuery.getAnalyzer())) {
+			matchPhrasePrefixQueryBuilder.analyzer(matchQuery.getAnalyzer());
+		}
+
+		if (matchQuery.getMaxExpansions() != null) {
+			matchPhrasePrefixQueryBuilder.maxExpansions(
+				matchQuery.getMaxExpansions());
+		}
+
+		if (matchQuery.getSlop() != null) {
+			matchPhrasePrefixQueryBuilder.slop(matchQuery.getSlop());
+		}
+
+		return matchPhrasePrefixQueryBuilder;
+	}
+
+	private QueryBuilder _translateMatchPhraseQuery(
+		String field, String value, MatchQuery matchQuery) {
+
+		MatchPhraseQueryBuilder matchPhraseQueryBuilder =
+			QueryBuilders.matchPhraseQuery(field, value);
+
+		if (Validator.isNotNull(matchQuery.getAnalyzer())) {
+			matchPhraseQueryBuilder.analyzer(matchQuery.getAnalyzer());
+		}
+
+		if (matchQuery.getSlop() != null) {
+			matchPhraseQueryBuilder.slop(matchQuery.getSlop());
+		}
+
+		return matchPhraseQueryBuilder;
+	}
+
+	private QueryBuilder _translateMatchQuery(
+		String field, Object value, MatchQuery matchQuery) {
+
+		MatchQueryBuilder matchQueryBuilder = QueryBuilders.matchQuery(
+			field, value);
+
+		if (Validator.isNotNull(matchQuery.getAnalyzer())) {
+			matchQueryBuilder.analyzer(matchQuery.getAnalyzer());
+		}
+
+		if (matchQuery.getCutOffFrequency() != null) {
+			matchQueryBuilder.cutoffFrequency(matchQuery.getCutOffFrequency());
+		}
+
+		if (matchQuery.getFuzziness() != null) {
+			matchQueryBuilder.fuzziness(
+				Fuzziness.build(matchQuery.getFuzziness()));
+		}
+
+		if (matchQuery.getFuzzyRewriteMethod() != null) {
+			String matchQueryFuzzyRewrite = MatchQueryTranslatorUtil.translate(
+				matchQuery.getFuzzyRewriteMethod());
+
+			matchQueryBuilder.fuzzyRewrite(matchQueryFuzzyRewrite);
+		}
+
+		if (matchQuery.getMaxExpansions() != null) {
+			matchQueryBuilder.maxExpansions(matchQuery.getMaxExpansions());
+		}
+
+		if (Validator.isNotNull(matchQuery.getMinShouldMatch())) {
+			matchQueryBuilder.minimumShouldMatch(
+				matchQuery.getMinShouldMatch());
+		}
+
+		if (matchQuery.getOperator() != null) {
+			org.elasticsearch.index.query.Operator operator =
+				MatchQueryTranslatorUtil.translate(matchQuery.getOperator());
+
+			matchQueryBuilder.operator(operator);
+		}
+
+		if (matchQuery.getPrefixLength() != null) {
+			matchQueryBuilder.prefixLength(matchQuery.getPrefixLength());
+		}
+
+		if (matchQuery.getZeroTermsQuery() != null) {
+			ZeroTermsQueryOption zeroTermsQueryOption =
+				MatchQueryTranslatorUtil.translate(
+					matchQuery.getZeroTermsQuery());
+
+			matchQueryBuilder.zeroTermsQuery(zeroTermsQueryOption);
+		}
+
+		if (matchQuery.isFuzzyTranspositions() != null) {
+			matchQueryBuilder.fuzzyTranspositions(
+				matchQuery.isFuzzyTranspositions());
+		}
+
+		if (matchQuery.isLenient() != null) {
+			matchQueryBuilder.lenient(matchQuery.isLenient());
+		}
+
+		return matchQueryBuilder;
+	}
+
 	private GeoShapeQueryBuilder _translateQuery(GeoShapeQuery geoShapeQuery) {
 		if (geoShapeQuery.getIndexedShapeId() != null) {
 			GeoShapeQueryBuilder geoShapeQueryBuilder =
@@ -896,9 +1042,6 @@ public class ElasticsearchQueryTranslator
 
 	@Reference
 	private MatchAllQueryTranslator _matchAllQueryTranslator;
-
-	@Reference
-	private MatchQueryTranslator _matchQueryTranslator;
 
 	@Reference
 	private MoreLikeThisQueryTranslator _moreLikeThisQueryTranslator;
